Include requests with 304 status codes
As pointed out by Bob West (thanks Bob!) 304s are perfectly valid
pageviews - they're just for content that is locally cached
and has not changed server-side since it was cached. We should
include those pageviews.

This patch does so, and also expands the test resources to
include an example with a 304 HTTP status, to prevent regressions.

Change-Id: Ic786037360eeae33726ffea3293d2c4dbb96f10a