Add new UDF to determine client IP given source IP and XFF header
ClientIpUDF, the new UDF being introduced, uses the IpUtil class added to
refinery-core. IpUtil maintains a list of trusted proxies against which the
source IP and the addresses from X-Forwarded-For header are compared in order
to determine the client IP. It uses the "spring-security-web" package from the
Spring project for its needs. Hence, we have new dependencies introduced in
both the main pom.xml file as well as in refinery-core/pom.xml.

Change-Id: I65a7914037fb097faabbaa5d817cb0b0838c12a8