De-static-everything
Why else would I call the branch 'degaussing'?

This is an architectural change to do something we've been talking
about for a while - namely, moving the generic functions for
string or regex comparisons out of specific classes, and then
making the classes dynamic rather than static, because not
everything needs to be static.

This patch:

1. Creates a "Utilities" class containing the regex and string
checker methods that previously lived in the Pageviews class. It
doesn't have unit tests because if those break, everything above it
also breaks.
2. Makes everything non-static (except for the Utilities, because
they really should be, and the referer...thing, because I took
one look at it and decided I probably shouldn't risk touching it).
3. Amends all the UDFs and unit tests so that they instantiate
the classes to call their public methods, rather than assuming
they're static methods that can just be called as-is.

Ackbar is back.

Change-Id: I7c5928a7f8a9bd6fe7eae0ecc546ff2122b862d0