Update app Pageview detection and expose it
Marcel needs the app Pageview detection logic to be exposed
and callable as a UDF. At the same time I discovered some issues
with the app formatting that makes pageview detection look way different
on different platforms.

This patch implements the latest logic for app Pageview detection (as
we understand it), exposes isAppPageview as a public method, and adds
a set of associated unit tests and an associated UDF. Further work
will be necessary to integrate with Apps' upcoming changes
(check your mailing lists).

Change-Id: I16b19e21b0b2cce9580bf6af7c97fc899b47d444