Add Apps session metrics job
Uses Twitter's algebird to compute percentiles

Removed geometric mean computation (not doable)
This job runs for 1 day of data with 6 executors and
and 2g per executor in 6 minutes if after we filter data
we lower parallelization 1 order of magnitude to do computations.

Input to the job is an output directory where to persist the data
and a date. Job appends data to file.

Bug: T86535
Change-Id: Ibd9bc45179ace3f74fe1bf1ac2f42de7867db90e