Add Search-centric UDFs
Two search-centric UDFs. One, SearchClassifier, identifies the type of API
search a request represents. The other, IsSearch, wraps around SearchClassifer
to provide a simple boolean response to "was this a search?"

One thing worth noting is that the UDFs themselves do not have unit tests...because
they're UDFs. They consist exclusively of a wrapper around the underlying Java, which does.
If we have input from Hive that passes /those/ unit tests it's only because it's unexpected
input that entirely duplicative UDF unit tests wouldn't have caught.

Change-Id: I2f785c873d49a0c0c51fa1c7492ea7d7a89c2d1a