Add support for custom timestamp and schema rev id in avro message decoders
AvroBinaryMessageDecoder and AvroJsonMessageDecoder now supports custom
timestamp values and format set in:
* camus.message.timestamp.field: field name (default to "timestamp")
* camus.message.timestamp.format: field value format

Formats supported are :
* unix_milliseconds
* unix_seconds
* ISO-8601
* Any format supported by java SimpleDateFormat

This is needed for CirrusSearchRequestSet to allow the following settings:
camus.message.timestamp.field: ts
camus.message.timestamp.format: unix_seconds

This adds also support for schema rev in the kafka message.
* AvroBinary the schema id is encoded in first bytes of the kafka message.
* AvroJson the schema id is loaded from a json property:
  - schemaID is the default name of the property
  - camus.message.schema.id.field can be set to specify a custom property

A backward compat property is available to set the writerSchema for producers
that do provide the schema rev id in the message:
camus.message.schema.default=123
This is done for backwards compatibility and will be used
when no schema is found in the message.

This backward compatibility code is 10 lines in each decoder.
It will just ease the deployment process because we can deploy without stopping the stream.
We have uploaded another patch that drops backward compat code: Ia605d22.

By default only the writerSchema will be used. We will rely on hive to do the
proper schema conversion.

It is also possible to force a targetSchema if
org.wikimedia.analytics.refinery.camus.schemaregistry.KafkaTopicSchemaRegistry
is used as the schema registry. The target schema can be forced by setting:
org.wikimedia.analytics.schemas.${SchemaName}.latestRev = 123

KafkaTopicSchemaRegistry has been refactored to allow easier plug-in
of other schema sources. So far schema lookups are going to disk.

Bug: T117873
Change-Id: Ie575f47110621972f820d49191b23510048afb37