Add UDF for network origin
Add a new UDF that can be used to classify the network origin of an IP
address. This broadly partitions all of the Internet into "internal"
(originating from an IP owned/leased by the Wikimedia Foundation),
"external" (originating from an IP not owned/leased by the Wikimedia
Foundation), and "labs" (subset of "internal" that has been allocated to
the Wikimedia Labs project).

The UDF is a wrapper around a new 'getNetworkOrigin' IpUtil method. That
method uses the existing trusted proxies list to determine "internal"
membership and a collection of IPv4 and IPv6 CIDR ranges copied from the
operations/puppet.git '$all_network_subnets' global variable to determine
"labs" membership.

Bug: T118592
Change-Id: Ib93584903fd9e2d0c6ab753440cb1e410d426e37