Create and cleanup temp dirs in TestCamusPartitionChecker.scala
Add a proper temp directory factory for use in tests that modify the file
system rather than relying on a shared testing directory and cleanup at the
start of the test. Each test invocation will securely create a temporary
directory and hang on to it in a class variable so that it can be cleaned
up by afterEach().

Bug: T119101
Change-Id: I399aff9f3a0b928cc0ab0c2e78278526778a0db7