Implement ArraySum UDF
For the CirrusSearchRequestSet I needed the ability to sum an array of
int's, along with the ability to ignore our sigil value due to issues
supporting [int,null] unions. This provides an initial simple
implementation of that sum.

This goes one step further, and sums an array of numeric values. I wasn't
sure what to do about overflows so those are unhandled, the UDF returns
the same type as the elements of the array it recieved. This is a bit slower
due to the use of BigDecimal, but my understanding of hive is that it is
almost always IO limited and some extra CPU usage will have minimal impact.

Change-Id: I56bfa6ce6c4ea383dacbb563a8b2dd96260a88a4