Functions for categorizing queries.
This adds functionality to deconstruct queries into features we
would be interested in -- e.g. whether the query has logical
operators like AND and NOT, how many double quotes there are, &
if there are wildcards.

For each query, the UDF produces an array of features such as:
[has AND, is insource, has even double quotes], [is simple],
[is fuzzy search], and [has wildcard, ends with ?]. The analyst
can then separate the features & perform various aggregations,
such as counting the number of queries having a particular
feature or combination of features.

The latest (final?)  revision (as of 2016-02-24) has the following
approximate performance statistics:
* 1.4-1.7 hrs to process 101-117 million queries per day
  (8 Feb 2016 - 17 Feb 2016) averaging to 0.050-0.054 ms/query
* 16.4 hrs to process 1.118 billion queries (0.529 ms/query)
* 1 hour of queries (median = 4.56 million) takes about 4 minutes

Bug: T118218
Change-Id: I6d83fb533fb8063f8887d3ab061a6be04857e652