Drop support for message without rev id in avro decoders and make latestRev mandatory
Drop support for legacy mediawiki kafka messages:
Avro message decoders accept a property that allows to use a default
schema to use when no rev id was provided in the kafka message.
This patch removes this backward compat code.

Make org.wikimedia.analytics.schemas.SchemaName.latestRev mandatory:
The previous patch allowed camus to run without a latest schema (using whatever
schema revision was found in the first kafka message). Unfortunately this does
not work and cause issues when the schema rev id change within the same camus
run. Camus will instantiate a DatumWriter and generate a file on the first
message and insert schema metadata in the file header, but if another message
is consumed within the same run the DatumWriter and file headers won't be
updated. This causes the whole run to fail (see T121483).

Change-Id: Ia605d22e65f6ac5f8f0e7fc851180b055ec29518