Upgrade to latest UA-Parser version
This patch upgrades refinery source to use the latest version of
UA-Parser (1.3.0-wmf3) - this update includes some upgrades to uap-core
yaml definitions and bug fixes to uap-java. Tests that break in refinery-
source due to the upgrade are also fixed - the breaking tests were due
to the new ua-parser differentiating between Mobile Safari in iOS and
Mobile Safari UIWebview in iOS apps.

Bug: T130399
Change-Id: I83d2686037b7db9d2ba2130cb7468a7c3c8e40e8