Evaluating Pageview tagging only for apps requests
Requests that come tagged with pageview=1 in x-analytics
header are considered pageviews if they are coming from an app
client. We look at the value of this header once
we have verified other attributes of the request that are needed
for it to be consider a pageview, like http status.

Bug: T128612

Change-Id: I5b74b9a2c9856fcbacda521fe5fb388e28307d40