Correct CqlRecordWriter in cassandra module
CqlRecordWriter was endlessly looping (generating mapreduce timeout) when
small-enough data was to be inserted into cassandra. The multithreading
approach in conjunction to the small data was leading to an Interrupt
signal being sent to the Cassandra Driver during connection initialisation,
leading to reconnection failure. Approach taken is to regularly check for
end-of-job condition instead of blocking and waiting for being Interrupted.
If not perfectly clean, this method at least prevent the observed failure.

Change-Id: If851c3e97f5c5ff083fb9775c8665ff35355f2d1