Match title & ns using x_analytics header & get all agent_types
The previous version of the job matches pageviews based
on the pagetitle "Special:AboutTopic".
This was an issue on sites where the namespace name
had been translated or an alias was used.
The ns ID is now added to the analytics header to avoid
this issue so -1 is matched directly for Special:

The previous version would also run into issues in the
future when aliases for the special page name would
be added.
Thus the raw name of the special page is now added to
the analytics header when viewed, and that is now used in
this patch.

The previous version also only added data about 'users'
to graphite, but a request for spiders has also been made
this the spiders are now also added.

x_analytics element docs were added to
https://wikitech.wikimedia.org/wiki/X-Analytics
in
https://wikitech.wikimedia.org/w/index.php?diff=746636&oldid=710297

A sample run of this query for 1 day:
 - Cumulative CPU: 66844.7 sec
 - Total MapReduce CPU Time Spent: 18 hours 34 minutes 4 seconds
 - Time taken: 442.308 seconds

Bug: T138500
Depends-On: Iaecfbae4c04c12472346b7f885b8a6e458bf47c4
Depends-On: I26c7b5ea6294b2403b89bd91869d2f18275c9917
Change-Id: I3848ee6136d2f7737d45071ee8528864c34299b4