Make camus partition checker more resilient
In previous version camus partition checker was failing in the
middle of a run if an error occurred on a topic (missing topic, index not moving...).
This was preventing work that should have happened after the failure to be done.
In this version if an error occurs, an appropriate message is logged but execution
is not stopped.

Bug: 144716
Change-Id: I2a15c16d21643dc7ec6658e9e1c5c524a524dc46