Lucene Stemmer UDF
Implements a stemmer for text using lucene for a variety of languages.
This is needed for some processing of search queries, to group together
user input that is approximately the same but not exactly. It may be useful
for other use cases that want to normalize strings as well.

Via testing in Spark and Hive we found we need different code
to initialize UDF in each.

Bug: T148811

Change-Id: I458e7ac724fefe813732b48fcfcef4728359fca9