UDF for extracting primary full text search request
CirrusSearch logs contain each individual request made from mediawiki to
elasticsearch. Often times we need to extract only the primary request
to do analysis on. It's quite a pain to do in HQL, so this UDF handles
it for us.

Adds a helper class for dealing with the structs, as I have the feeling
there will be more use cases for UDF's that directly handle the list of
structs.

Bug: T162054
Change-Id: I67d5f0e7674f970b353ab5992fec1431f4592256