[cirrus] Distinguish morelike vs fulltext api search requests
Since mobile has enabled the RelatedArticle feature on 90%
of the taffic on enwiki it's hard to have an idea of the real
fulltext API usage.

This patch moves the patterns into a separate ENUM that can be
augmented if/when we augment the API. (Although it's okay to go
back to the previous if-else logic version of this.)

NOTE: this is just a suggestion and will require a followup patch
in our dashboards to take into account this new search api type.

ALSO: this fixes an incorrect deprecation message that pointed
users to the wrong UDF for categorizing search requests.

Change-Id: I7cc701bdc985323fc1fb0a74cc72c67fcdc65b10