UDF to tag requests
UDF that understands webrequest data and can classify
requests into types like "portal", "wikidata" and others. It uses a
set of classes to look at requests and tag them
as belonging to a type, a request can belong
to several types.

To create a Tagger you must implement the tagger interface
and annotate your class with @Tag

Tested with guava 12.0

Bug: T164021
Change-Id: I725e130431f3a869864275aa41af479fef9f157c