Refactor PageviewDefinition to add RedirectToPageviewUDF
In order to count uniques on global domains, we need to look
at redirect-to-pageviews in addition to pageviews. The function
checks that the request is a pageview, except for the http status
code and response content-type. The former should be either 301,
302 or 307 (instead of 200 or 304 for natural pageviews), and the
latter also accepts non-set content-types.

Refactor PageviewDefinition code for easier readability
so we can add IsRedirectToPageviewUDF w/o much changes to
Pageview code

Bug: T143928

Change-Id: I9e381da8f49d18c521590206d2d96bdcd25dd3aa