Add tranquility to the banner streaming job
In order to productionise druid pseudo-realtime ingestion
we want to use tranquility. We used to have standalone processes
reading from Kafka and inserting in druid, but is feels easier
to have tranquility running as part of the streaming
transformation job.
This patch also provides parameters and error handling making
the streaming job much more resilient.
Finally, due to dependency issues, we need to make this job run on
spark 2.1.1, and in order to do so we put the job in a new temporary
package to which we'll iteratively move spark jobs when moving to
new spark version.

Bug: T168550, T169101

Change-Id: I22b0b641ad2ad2f1c88888332b03373abe3e0d00