Add _REFINE_FAILED failure flag and skip refinement if it exists
Since JsonRefine selects targets well into the past, we need to not only
skip targets that previously succeeded, but also those that failed.
Keeping failure state in the _REFINE_FAILED flag allows us to do this.

_REFINE_FAILED works the same as the _REFINED flag, in that it stores
the mtime of the input target.  If the input target's mtime changes,
it will be selected for (re) refinement, even if the _REFINE_FAILED
flag exists.

Change-Id: I69e344a7ea24631e446436b32623223f06123860