Add dataframe conversion to new schema function
In order to prevent reading and transforming the json data twice,
we need to convert the already read and transformed dataframe to
match the newly computed schema. We know the new schema is an
extension of the old one with reordering, so the function is
about reordering fields and setting new null values.

Change-Id: I6d569fb7942f02852580e716340da20bd56661f6