Get smart and hacky about compatible type casting
Ran into a case in production where a JSON number of
9223372036854776000 was causing an entire JSON dataset
hour to be inferred with a string field, even though all
other records had a totally valid long values.  Since
String and Long were not compatible, this caused the entire
hour to fail refinement.

This patch adds some hacky code to do some type casting if we can.
In the case above, this would end up casting the bad value to Null,
leaving the rest of the record and dataset intact! :o Amazing! This
would also handle the case where someone somehow gets a string
number, e.g. "12345" into the data where an integer should be.

This also adds a case for DecimalType -> LongType casting,
as Spark 2 will infer the 9223372036854776000 value as a
DecimalType(20,0) (which wraps BigDecimal).  Nice Spark 2!
But in this case, the proper type is still Long, so we
are fine with nulling out the 9223372036854776000 in Spark 2 as well.

Bug: T189332
Change-Id: Ie7e64e1506e36ddff1963b71cd76efc12e380e32