Update MediawikiHistory-denormalize
In the last commit provided for performance optimization,
some bits were left out: removing the by-year grouping of
per-user and per-page data. Now each page or user history
in terms of revision is worked entirely in a single partition,
preventing data duplication in case of joining with user/page
events, and also preventing shuffling steps for per-user and
per-page metrics, as his can now happen when joining with
page/user events.
Finally spark UI parameters have been set so that less data
is stored in the driver, preventing it to fail with OOM.

Change-Id: I3e87c9244da5dffa828ebbd641879e0476e31e0d