Update mediawiki-user-history empty-registration
We were keeping in the dataset users having undefined registration
date. Now that we want to index the data from the parquet files,
these null timestamp are not acceptable anymore.
This patch logs those rows as errors instead of keeping them in
the dataset.
We also are sure the discarded rows have no edit activity since
the registration-date is set to
MIN(DB-Registration-date, first-edit-date).
One solution for later (documented as a TODO in the file) is to
extrapolate the registration date from the sequence of IDs and
dates.

Bug: T201620
Change-Id: I419a8d7fdbc1af9b030da0fff5a368ff46207f07