mediawiki_history_checker job running for the wrong set of wikis
After a mediawiki history snapshot is created we run a set of automatic
steps to asses quality of snapshots data, these validation steps should
be run in the top  wikis in terms of events as smaller wikis might
have too much variability among snapshots and thus comparing them is
not possible.

We need to compute the wikis we are running events for
by using "SORT BY" instead of "ORDER BY" in spark sql queries.

"SORT BY" sorts inside partitions while "ORDER BY" does a
full sort. Since the job uses multiple partitions, the result
of "SORT BY" was not a  global sort.

Change-Id: I2e0f102256a391ffdf31fc4e8508042580ee47fd