Upgrade profig lib to 2.3.3 after bug fix upstream
See https://github.com/outr/profig/pull/25

The changes on the implicit boolean converter are necessary
to avoid a collision in implicit definitions between
the converter defined within ConfigHelper and the new
version of profig library (in the Conversions module).
As the new profig code already includes such converter,
this change removes the one from ConfigHelper.

Bug: T202429
Change-Id: I2cb32194099dbe7eacfee68857e1a5612de68f49