Update mediawiki-history comment and actor joins
Joining to comment and actor table in mediawiki-history has
lead to issues due to skewed data: the number of revisions sharing
the same comment_id can be big (but for a small number of comments),
and for the moment as the actor table is not used, most actor_ids are
NULL. This skew impacts distributed-joins as they involve spreading
the data on workers by join-key (here comment_id or actor_id), making
some worker overwhelmed with the amount of data to work.
We solve the problem with two different tricks:
 - We colaesce the NULL actor_ids to a negative value (not joining
any real data), enforcing spreading the NULLS among workers.
 - We 'explode' the comment_id shared by many revisions by precomputing
the ones that need it, defining a map-side join function allowing to
know for each comment_id of they need to be exploded or not (and by how
many), and defining a map-side function duplicating the comment table
rows that need to be joint through explosion.

Bug: T210543
Change-Id: I68675186a44768c5f24ceb4842047c60ee689c73