Refactor mediawiki-page-history computation
This patch is a full refactor of mediawiki-page-history code.
This refactor involves using pageId when it's available,
and pageTitle when pageId is not available, to join events.
This patch also changes the organization of the code trying to
make it more readable and maintainable despite a more complex
algorithm.
A bunch of tests have also been added.
The result is better, but still not perfect when pageId are not
available.

Bug: T190434
Change-Id: I5d3d26bdf02151ca4e466f29af3c6e1e528394ce