Add SparkSchemaLoader capabilities to Refine and RefineTarget
This allows Spark schemas to be loaded for a given RefineTarget when
and used whewn reading it in as a Spart DataFrame.

The EventSparkSchemaLoader implementation uses Event(Logging)SchemaLoaders from
refinery-core to get the JSONSchema of event datasets and then uses
refinery-spark JsonSchemaConverter to convert to Spark StructTypes.

This refactors some parts of RefineTarget to allow for reuse of inputFormat
as a lazy val, and to use a default SparkSchemaLoader that always returns None.

Also add logging to Spark JsonSchemaConverter, properly use long for integers,
and support JSONSchema Draft 3 style 'required'.

Bug: T215442
Change-Id: I9a492cb55ede5eeb7db420d0fdca5194a6b78613