Reject invalid Page titles
Titles are checked for length and for illegal characters, any corrupt
titles are rejected and set to the unknown marker, "-".  We don't
perform any of the more sophisticated checks like in MediaWikiTitleCodec.

This sanitization is necessary to prevent illegal titles from leaking
through to the pageview dumps, where they can break parsing for example
when the string contains a newline.

Bug: T144100
Bug: T152628
Change-Id: I9976e910a790fcedd7ccdb1cea83a31006b2de3b