Handle dropping of user fields in labs views
The actor and comment tables have replaced literal string values in
logging, revision, archive, and other tables.  There was a transition
period where both the new ids pointing to the actor and comment table
and the old literal value columns were in the tables (or views when querying
cloud dbs).  Now, the literal columns have been removed.  This means our
job has to remove references to them.  This is a fairly manual and
error-prone activity because we don't have tests for integration with
outside data, and because we have to manually index into rows of data
returned by queries.

Also note that I tried to add comments or explicit type declarations
wherever possible because one of the bugs I found while refactoring was
not finding all the case class copy calls that needed to be updated with
the new field.  This should help finding Page and User Events and States
going forward, and it seems like a good code practice to introduce.

TODO: running history on last month's snapshot (doing so now)

Bug: T224134
Change-Id: Ic2b5bc8e574755dd197ca392dfa09dde2c648488