Add UDF to get wiki project from referer string
At the moment the only way we have to obtain a project dimension in
the mediacounts dataset is by parsing the referer string and save which
wiki site the file was requested from.

This change also refactors some stateless functions that were in the
PageviewDefinition class into the Webrequest class.

Bug: T228151
Change-Id: I93a93c154fd792a21b92cf3f731afd46c669f0ea