Remove BannerImpressions streaming job and deps
The BannerImpressions streaming job is not maintained anymore nor is
Tranquility to send realtime events to Druid. Druid is now providing
a way to stream data directly from kafka through Supervisors.
This patch removes the classes and maven dependencies.
In addition to dead-code being removed, dependency removal leads
to a 30% decrease of the size of the refinery-job shaded-jar.

Bug: T245151
Change-Id: I371100ff78607925cae442c932ed3ab48ddf79de