Support multiple possible schema base URIs in EventSchemaLoader
This allows EventSchemaLoader to be instantiated with multiple
base schema URIs in which to look for event schemas.  The first
schema found will be returned.

Also add the ability to load the 'latest' event schema instead
of the $schema URI specified by the event.

EventSparkSchemaLoader for Refine now has a loadLatest
parameter (default true) to load the latest event schema.

Bug: T240985

Change-Id: I8b7692d27e9d7de8a9728e95ef44ad54b27c621d