Introduce RefineFailuresChecker.scala
Fork Refine.scala to create a new util script aimed to
check targets with _REFINED_FAILED flag. Based on Andrew's code:
https://gist.github.com/ottomata/6841ee47f8b4b86081856a1ff636d38f

The idea is to run this from an-coord1001 daily to check the past 24
hours' targets, and alarm in case one or more are found. This should
help people reviewing a burst of refine alarm emails.

Bug: T240230
Change-Id: I0ec7c45671361232be3f500e8a63c6b9526d8119