Add ActorSignatureGenerator and GetActorSignatureUDF
The UDF encapsulates the generation of a hash serving as an
'actor' signature for the purpose of labeling automatted traffic.

It reproduces the following hive clause:
md5(concat(ip, substr(user_agent,0,200), accept_language, uri_host,
           COALESCE(x_analytics_map['wmfuuid'],
                    parse_url(concat('http://bla.org/woo/', uri_query), 'QUERY', 'appInstallID'),
                    '')))

This UDF is usefull to prevent having to rewrite the function manually in every place we need it.

Bug: T247342
Change-Id: I4342f4bf96bcca8c2dd921b9a0619fced8659dbc