Add EvolveHiveTable CLI tool to manually evolve Hive tables from JSONSchemas
This works just like Refine does, but allows us to manually evolve a table
rather than relying on new event data to do so as part of Refine.

This will be used as part of the legacy EventLogging event migration
to EventGate so that we can incrementally migrate EventLogging legacy
streams using the on metawiki schemas.  We will pre-evolve the
legacy Hive tables using the MEP schema, but continue to refine
them with the metawiki schema untnil the events have fully
moved to EventGate.

Bug: T238230

Change-Id: Ic04c105e939b4633f2bd63925ec5ca93657fa326