Remove filter_allowed_domains from common event_transforms
This function should not be applied to events that we know are emitted
from internal WMF clients.  Only external EventLogging based
events need to have this filter applied.

This was causing wdqs_external_sparql_query to have 0 events, since
query.wikidata.org is not a 'wikimedia domain' according to Webrequest.

In the future, we should probably add a new column during Refinement,
  is_wikimedia_source_domain
To indicate that an event passes the allowed domain whitelist instead of
filtering out ahead of time.

Change-Id: Ib1f98a012a9cac26db82beeeada4cb667197ff6d