Pull plugins from Archiva instead of Maven Central
During the last release, an error was raised due to http://repo.maven.org
not available via plain http, but only via https. We use central for plugin
download, meanwhile we do not (enabled=false) for artifacts.
This change disables plugin settings for maven central, forcing the usage
of Archiva.

Change-Id: I48e4656a9cbb51f87960b9fefe052454e11b2c28