Refine - Don't merge Hive schema by default when reading input data
This adds a new Refine option: merge_with_hive_schema_before_read
(default false) which controls this behavior.  We no longer
want to do this now that we have better controls over event
schema version evolution and compatilbility, but still need
to do it for schemas loaded for legacy EventLogging data from
metawiki.  This will require setting this flag to true
for the existent refine_eventlogging_analytics job before
deployment.

This moves the schema merge logic out of EventSparkSchemaLoader
and to the RefineTarget inputDataFrame function.  I think this
is a better place for it, as the only reason for merging at all
was for having a complete schema when reading input data.

I've also made the schema loaded for a RefineTarget
a public (lazy) val to more easily get its schema.

Bug: T255818
Change-Id: Ib6584fe526808a624d2885345a47d054c9d7df31