Spark JsonSchemaConverter - additionalProperties with schema is always a MapType
It would be useful to be able to write a JSONSchema with a MapType field
that still specifies some defined properties.  This would allow
schema owners to provide hints as to some keys that could be in
a Map, without having to explicitly define all keys.

JsonSchemaConverter will now always convert an object with aadditionalProperties
as an object (not a boolean) to a MapType.  In this case, properties will
be ignored.

Events schemas will use this to e.g. add required map keys.

Bug: T263466
Change-Id: I134d6f271693ce6079db4a65184ce8acc7950736