Update MediawikiXMLDumpsConverter repartitioning
The job applies a 'distinct' operation on the dataset to prevent
revision duplication, and this patch takes advantage of the step to
repartition the dataset into the configured number of partition
instead of doing it explictly as part of another step.
This means every processed XML file will go through one shuffle stage
instead of 2.

Bug: T263736
Change-Id: I161fa7d911919f4dfc050fb96d5702a99bd0cd81