Refine using PERMISSIVE mode and log more info about corrupt records
This patch adds a --corrupt_record_failure_threshold option, which
defaults to 1 record.  If the number of corrupt records is greater
than --corrupt_record_failure_threshold , the Refine of a dataset will fail
with an exception that includes information about the number of corrupt vs
total input records, and the first corrupt record as an example.  If the
number of corrupt records is less than --corrupt_record_failure_threshold,
an warning with the same information will be logged, but the job will
succeed.

This will make the chore of rerunning Refine jobs due to
malformed/corrupt records a little easier.  After reading
any failure alert and deciding that the job should be rerun,
run with --corrupt_record_failure_threshold=<# corrupt records + 1),
or run in DROPMALFORMED mode (which will have the same effect).

In a future more thought out patch, we should do something
useful with the corrupt records, like writing them
to a side output, or updating a Refine statistics table.

Bug: T266872
Change-Id: I350c701d04003bf7a19c632f695bd0581418c4c5