Fix failing sonar analysis due to JDK11 removing tools.jar
The Hadoop and Hive versions that we are currently using rely
the tools.jar provided by JDK <= 8. JDK11 removed this jar. Since
we are using JDK11 for sonar analysis, this breaks the analysis
and thus breaks the build. Sonar does not actually cares that
about that dependency, but needs a valid project.

Workaround: excludes tools.jar from dependencies when running on
JDK11. Note that a proper fix will be required when migrating to
JDK11 for real (probably upgrading Hadoop version).

To make the pom.xml more clear, dependencies requiring tools.jar
have all been declared in the <dependencyManagement/> section of
the parent pom. This way, the exclusion can be done in a single
place.

Change-Id: If082b03cc04c675ab76b96301b393f9b628fb1b6