Improve Refine failure report email
Include spark job name, as well as options useful
for reruning the failed RefineTargets.

Add a mapToDateTime function to HivePartition to
convert from a HivePartition ListMap to a DateTime.

Improve ConfigHelper DateTime parser to work with properly
formatted ISO-8601 date times, always in UTC.
This is safe to do, as we only use DateTimes for configs
when re-running jobs.

Change-Id: I8027e1bc7d08340aeaf82b4f4915280920eed5b4