Publish both shaded and unshaded artifacts.
Shaded jars are useful as an end artifact, but are problematic to reuse
as a dependency. So we should publish both shaded and unshaded artifacts.

This CR will rename the shaded artifact with a "-shaded" classifier, and
attach both the shaded and unshaded jar to the build.

The configuration of the maven-shade-plugin is moved to the parent pom
to reduce duplication and ensure that all modules are published in the
same way.

Bug: T217967
Depends-On: Icf7c4bc94ff53331dccbe7910f3dd77da044bae3
Change-Id: I30998356d497f0d32db89dd40e4c99e9289de84f