Refine - don't remove records during deduplication if ids are null
If an id used for dedupcliation is null, likely something is wrong
with the producer of data, and the records are not duplicates.
Don't remove records because of null ids.

This changes the behavior of deduplicate to only consider
the first possible non NULL id column, or use a fake uuid,
as the deduplication key.  In practice, this shouldn't matter.l
Legacy EventLogging events are the only ones that have
both uuid and meta.id fields, and either could be use as the deduplication
key on their own, and in practice only one of the two is set anyway.

Bug: T294361
Change-Id: If5f1e48e5ad7d88836570b383d5f882c20940877