Add SparkSQLNoCLIDriver job
This job is an updated copy of the SparkSQLNCLIDriver that is used
to start a spark-sql shell. The copy doesn't provide shell mode but
provides the same set of parameters the original job does to run
queries: directly as an argument, or reading a file, with variable
substitution. Not having the shell aspect allows to run the job in
cluster mode, which is needed when used with Airflow.

Bug: T297427
Change-Id: Ia090d95b8ebb4b8004605311280f4a9e6f716293