Integrate SparkSQLNoCLIDriver and HiveToCassandra
A new SparkSQLConfig trait facilitates implements parameters
parsing in scala and can easily be extended to be reused.

SparkSQLNoCLIDriver takes advantage of it and gives the
possiblity to get the result of the latest executed query.

These are used in HiveToCassandra to provide HQL running
capability as in SparkSQLNoCLIDriver mode, using the
result of the query to load cassandra.

Bug: T297934
Change-Id: I4a221de32beb6e160dd9662ce49497e8cffe5437