Make caching mechanisms thread ready
In some places, the code was using LinkedHashMap to cache the result of
repeated computations.

LinkedHashMap is not thread-safe. It was not a problem when called in the
context of Hive (Proc oriented). But it became a problem when used in the
context of Spark (Multi-threaded).

Without this fix, there is concurrency to fill those caches. And sometimes, the
result is some null returns from the cache (it could also trigger more serious
errors).

Here is a list of currently impacted jobs through UDFs:
- **refine** is alreday subject with this problem because it uses Spark
  https://github.com/wikimedia/analytics-refinery-source/blob/master/refinery-job/src/main/scala/org/wikimedia/analytics/refinery/job/refine/TransformFunctions.scala#L35
- hql/mediarequest/hourly.hql
- hql/session_length/daily.hql (through computed data in refine)
- ...?

Soon to be migrated Oozie jobs from Hive to Spark:
- oozie/interlanguage/daily/interlanguage_navigation.hql
- oozie/webrequest/load/refine_webrequest.hql
- oozie/pageview/actor/pageview_actor.hql
- oozie/webrequest/load/refine_webrequest.hql

The general bug is: T305386
The original task is: T300029

Bug: T300029
Change-Id: Idbe98dd66534b09a88c93be671c45bed24366de4