Update mediawiki-history page and user computation
This update makes the job result consistent when run over the same
data. There is a bug in the code about deleted-pages not being
processed correctly. The bug is documented here:
https://phabricator.wikimedia.org/T320860
Due to this original problem, and some other, the current version
of the job could compute different results for the the same original data.
This patches only fixes the consitency issue, not the bug described above.

Bug: T318589
Change-Id: Ic170e7059b1a581bb0601f888ba59e4ffbc73d94