Remove Guava from dependency
This patch replaces Guava caches with Caffeine. It includes 2 more
changes to make the UDFs caches works with Spark:

1/ About singletons (Webrequest & PageviewDefinition)

* This patch switch to thread-safe lazy instantiation of the singleton.

2/ About not making the singleton classes serializable

* Spark uses Kryo to serialize objects to be sent for task execution. Caffeine is
not Kryo friendly and doesn't support being serialized/deserialized out of the
box. This is not an issue as we use caches inside tasks with no cache
initialization, so making the cache transient is ok.

* Our singletons don't hold internal state that we'd like to keep at
serialization time (we hold cache data, but this is not to be serialized as
serialization happens before task launch). Therefore there is no real gain in
having singletons serializable and we make the singleton variables transient
when we need them.

Bug: T327072
Change-Id: I7a13b04213311f3eb569a0334e6855a657fe6757