Support snapshot partitioning in HiveToDruid and DataFrameToDruid
This change adds suport for snapshot-based partitioning schemes
in DataFrameToDruid.scala and HiveToDruid.scala modules.
So far they only supported year-month-day-hour-based ones.
This change also adds a couple convenient parameters to
those models, to be able to specify map memory resources
and specify the temporary directory to use.
Some of the code of HiveToDruid was duplicated in HivePartition,
I removed the first one in favor of the latter.

Bug: T324485
Change-Id: I31734a4279ccd1bb9319f706293bf09c53eac3fd